requirements.txt
Django==4.0.1
django-heroku==0.3.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
django-cors-headers==3.5.0
djangorestframework==3.13.1
gunicorn==20.0.4
whitenoise==5.2.0
drf-social-oauth2==1.2.0
Pillow==9.0.0    